refactor: extract TrackerExportSupport+CreateWithRetry.swift
Splits the createWithRetry state-machine extension (~113 lines, the
shared dup-safe create-with-retry loop used by both Jira and GitHub
providers per #502) out of TrackerExportSupport.swift into a
convention-named `+CreateWithRetry.swift` sidecar.

Byte-identical move of lines 94-206. Both files import Foundation.

TrackerExportSupport.swift: 206 → 93 lines. Tests: 667.
